third_party/WebKit/LayoutTests/http/tests/security/subresourceIntegrity/subresource-integrity-style-blocked.html

<!DOCTYPE html>
<html>
    <head>
        <title>Subresource integrity blocked style resources</title>
        <script src="../../resources/testharness.js"></script>

        <link rel="stylesheet" href="style-1-of-3.css" integrity="ni://sha256;BpfBw7ivV8q2jLiT13fxDYAe2tJllusRSZ273h2nFSE="></link>
        <link rel="stylesheet" href="style-2-of-3.css" integrity="ni://sha384;XVVXBGoYw6AJOh9J/Z8pBDMVVPfkBpngexkA7JqZu8d5GENND6TEIup/tA1v5GPr"></link>
        <link rel="stylesheet" href="style-3-of-3.css" integrity="ni://sha512;tbUPioKbVBplr0b1ucnWB57SJWt4x9dOE0Vy2mzCXvH3FepqDZ+07yMK81ytlg0MPaIrPAjcHqba5csorDWtKg=="></link>
    </head>
    <body>
        <script>
        if (window.testRunner)
            testRunner.dumpAsText();
        </script>
        <p id="p1">Test 1</p>
        <p id="p2">Test 2</p>
        <p id="p3">Test 3</p>

        <script>
        var color = window.getComputedStyle(document.querySelector('#p1')).color;
        assert_not_equals(color, "rgb(0, 128, 0)", "The p#1 element's text should not be green.");
        color = window.getComputedStyle(document.querySelector('#p2')).color;
        assert_not_equals(color, "rgb(0, 128, 0)", "The p#2 element's text should not be green.");
        color = window.getComputedStyle(document.querySelector('#p3')).color;
        assert_not_equals(color, "rgb(0, 128, 0)", "The p#3 element's text should not be green.");
        </script>
    </body>
</html>
